Split statics components
