+++ packages/ui/src/common/components/statistics/Statistics.tsx
@@ -1,0 +1,25 @@
+import React from 'react'
+import styled from 'styled-components'
+
+import { StatisticItem, StatisticItemProps } from './StatisticItem'
+
+interface StatisticsProps {
+  stats: Array<StatisticItemProps>
+}
+
+export const Statistics = ({ stats }: StatisticsProps) => {
+  return (
+    <Stats>
+      {stats.map(({ title, helperText, value, className }) => (
+        <StatisticItem title={title} helperText={helperText} value={value} className={className} />
+      ))}
+    </Stats>
+  )
+}
+
+const Stats = styled.ul`
+  display: flex;
+  width: 100%;
+  justify-items: flex-start;
+  margin-top: 8px;
+`
